+++ MatrixSDK/Data/RoomSummaryStore/CoreData/Models/MXUsersTrustLevelSummaryMO.swift
@@ -24,20 +24,20 @@ public class MXUsersTrustLevelSummaryMO: NSManagedObject {
         return NSFetchRequest<MXUsersTrustLevelSummaryMO>(entityName: entityName)
     }
 
-    @NSManaged public var s_usersCount: Int16
-    @NSManaged public var s_trustedUsersCount: Int16
-    @NSManaged public var s_devicesCount: Int16
-    @NSManaged public var s_trustedDevicesCount: Int16
+    @NSManaged public var s_usersCount: Int32
+    @NSManaged public var s_trustedUsersCount: Int32
+    @NSManaged public var s_devicesCount: Int32
+    @NSManaged public var s_trustedDevicesCount: Int32
     
     @discardableResult
     internal static func insert(roomUsersTrustLevelSummary usersTrustLevelSummary: MXUsersTrustLevelSummary,
                                 into moc: NSManagedObjectContext) -> MXUsersTrustLevelSummaryMO {
         let model = MXUsersTrustLevelSummaryMO(context: moc)
         
-        model.s_usersCount = Int16(usersTrustLevelSummary.trustedUsersProgress.totalUnitCount)
-        model.s_trustedUsersCount = Int16(usersTrustLevelSummary.trustedUsersProgress.completedUnitCount)
-        model.s_devicesCount = Int16(usersTrustLevelSummary.trustedDevicesProgress.totalUnitCount)
-        model.s_trustedDevicesCount = Int16(usersTrustLevelSummary.trustedDevicesProgress.completedUnitCount)
+        model.s_usersCount = Int32(usersTrustLevelSummary.trustedUsersProgress.totalUnitCount)
+        model.s_trustedUsersCount = Int32(usersTrustLevelSummary.trustedUsersProgress.completedUnitCount)
+        model.s_devicesCount = Int32(usersTrustLevelSummary.trustedDevicesProgress.totalUnitCount)
+        model.s_trustedDevicesCount = Int32(usersTrustLevelSummary.trustedDevicesProgress.completedUnitCount)
         
         return model
     }
